Fix getUint64 fallback param type

+++ config/get_viper.go
@@ -33,7 +33,7 @@ func getUint32(key string, fallback uint32) uint32 {
 	return viper.GetUint32(key)
 }
 
-func getUint64(key string, fallback uint32) uint64 {
+func getUint64(key string, fallback uint64) uint64 {
 	viper.SetDefault(key, fallback)
 	return viper.GetUint64(key)
 }
